[TEST] :  로또 객체의 책임과 무관한 테스트를 분리하여 관련 파일로 이동
