+++ __tests__/validator/BonusNumberValidator.test.js
@@ -1,5 +1,5 @@
-import { ERROR } from "../src/constants/message";
-import BonusNumberValidator from "../src/validation/BonusNumberValidator";
+import { ERROR } from "../../src/constants/message";
+import BonusNumberValidator from "../../src/validation/BonusNumberValidator";
 
 describe("보너스 번호 유효성 검사를 진행한다", () => {
   test("로또 번호중에 보너스 번호와 중복된 숫자가 있으면 오류를 발생시킨다.", () => {
