+++ __tests__/validator/PriceValidator.test.js
@@ -1,5 +1,5 @@
-import { ERROR } from "../src/constants/message.js";
-import PriceValidator from "../src/validation/PriceValidator.js";
+import { ERROR } from "../../src/constants/message.js";
+import PriceValidator from "../../src/validation/PriceValidator.js";
 
 describe("입력한 구매 금액에 대한 유효성 검사를 진행한다", () => {
   test("1000원 단위가 아닌 경우 에러를 발생시킨다.", () => {
